+++ docs/changelog.rst
@@ -13,17 +13,19 @@ The format is based on `Keep a Changelog`_, and this project adheres to `Semanti
 - Function to query the version of SimpleBLE at runtime.
 - (Python) Missing API from SimpleBLE::Characteristic.
 
-
 **Changed**
 
 - (MacOS) Main adapter address is now hardcoded to allow caching based on adapter address. *(Thanks BlissChapman!)*
 - (Python) Release GIL when calling ``Peripheral.write_request`` and ``Peripheral.write_command``.
+- (MacOS) Operations are now event-based.
+- (MacOS) OperationFailed exception now contains the error message provided by the OS.
 
 **Fixed**
 
 - (MacOS) Remove unnecessary timeout during service discovery. *(Thanks BlissChapman!)*
 - (MacOS) Return correct list of devices when scanning. *(Thanks roozmahdavian!)*
 - (MacOS) Remove unnecessary timeout during characteristic notification. *(Thanks BlissChapman!)*
+- (MacOS) Remove unnecessary timeout during operations on characteristics.
 - (Windows) Failed connection attempt would not trigger an exception. *(Thanks eriklins!)*